fix: SPDX validation failure when no packages are present (#349)
When no packages (e.g. CyTRICS SBOM systems or software entries that
contain other software entries) are present, SPDX output fails with
a validation message saying a `SPDXRef-DOCUMENT DESCRIBES ...` or
`... DESCRIBED_BY SPDXRef-DOCUMENT` relationship must be present.

To fix this, a relationship saying NOASSERTION is added if no
relationships to the SPDX Document are created for "packages".

Resolves #172