feat: remove limitation on gzip file type recognition (#373)
For historical reasons, gzipped files were only identified as if they had a .tar.gz
or .cab.gz extensions to limit the size of the generated SBOM (e.g. .html.gz files
would not be included in an SBOM). With the excludeFileExts options, this limit is
no longer necessary.

This change also required improving the error handling to detect compressed Docker
archives, to avoid errors if the gzipped file being checked is not a compressed
tar file.