feat: support plugin hints for software entry fields (#374)
Adds a new optional argument for extract_file_info hooks to use called
software_field_hints. Plugins can append tuples to the list that provide
a suggested value for various fields in the software entry (such as name,
version, etc), and some slightly arbitrary measure of how confident they
are in the accuracy of the information.

If the fields already have a value set, the existing value takes priority
(e.g. a user provided a starting point SBOM or an earlier run already picked
a value). Vendor is a special case where any newly detected vendors will just
be appended to the list of vendors.

Resolves #8