feat: Add fast pattern matching using AhoCorasick searching (#388)
Add support for quickly locating places in a file that might match a regular expression. It results in the runtime remaining roughly constant as the number of patterns increases, enabling matching of 100k+ patterns in under a minute in situations that would take over 30 minutes when checking each regex individually. For larger files, it will result in the time taken being 1/n of the current time, where n is the number of patterns being searched for.

* Add fixed regex literal extraction code, by parsing regex into a tree and searching for literals with some support branches/subpatterns/character classes that add multiple possibilities up to a certain threshold
* Add aho-corasick automaton builder to quickly find "needles in a haystack" that match the identified regex literals
* Use aho-corasick automaton in native lib detection to quickly search file contents for places that are most likely to match the regular expression, and then check if the regular expression matches the contents in the part of the file that was identified

A minimum prefix length of 3 is used to avoid an excessive number of matches against extremely common characters (such as 0-9), and the maximum number of prefixes computed per pattern is capped at 10 to avoid needing to construct an extremely large automaton (changing this value is a balance between higher likelihood of matching the regex at identified locations vs needing to check more places for a regex match -- 5-15 seemed like a bit of a sweet spot).

Resolves #306 - performance improvement seems large enough that hyperscan likely won't yield a similar order of magnitude speedup over this approach. Could still explore that option in the future, but this is looking pretty good and is pure python.